move place bid networking into a model
